Changed tests Changed polygon-polygon collision

+++ ProjetoPOO/src/Point.java
@@ -55,7 +55,7 @@ public class Point
      */
     @Override
     public String toString() {
-        return  "(" + x + "," + y + ")";
+        return  "(" + (int)x + "," + (int)y + ")";
     }
 
     /**
